+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CLAide Changelog
 
+## Master
+
+##### Enhancements
+
+* None.  
+
+##### Bug Fixes
+
+* None.  
+
+
 ## 1.0.0.beta.3 (2016-03-15)
 
 ##### Enhancements