+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CLAide Changelog
 
-## Master
+## 1.0.0.beta.3 (2016-03-15)
 
 ##### Enhancements